--- src.com.cod.mvc/controller/Controler.java
@@ -1,35 +1,0 @@
-package controller;
-
-import model.Model;
-
-public class Controler {
-    private final Model miModel = new Model();
-
-    public Controler(Model miModel) {
-
-        ObserverVelocidad ObVelocida = new ObserverVelocidad();
-        miModel.addObserver(ObVelocida);
-
-        ObserverLimite ObLimite = new ObserverLimite();
-        miModel.addObserver(ObLimite);
-
-    }
-
-    /**
-     * Crea un coche
-     * @param modelo del coche
-     * @param matricula por la que buscar el coche
-     */
-    public void crearCoche(String modelo, String matricula){
-        miModel.crearCoche(matricula,modelo);
-    }
-
-    /**
-     * Cambiar la velocidad de un coche
-     * @param matricula por la que buscar el coche
-     * @param velocidad nueva
-     */
-    public void cambiarVelocidad(String matricula, Integer velocidad){
-        miModel.cambiarVelocidad(matricula, velocidad);
-    }
-}
